backend/pyproject.toml
[tool.poetry]
name = "teacher-ye-answer"
version = "0.1.0"
description = ""
authors = ["linyh <l@likyh.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.6"
django-ninja = "0.22.2"
openai = "^0.28.1"
pymysql = "1.0.3"
python-dotenv = "1.0.0"
requests = "2.31.0"
mistune = "3.0.2"
langchain = "^0.0.325"
psycopg2 = "^2.9.6"
psycopg = "^3.1.9"


[tool.poetry.group.dev.dependencies]
psycopg2-binary = "^2.9.6"
psycopg-binary = "^3.1.9"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

#设置poetry包管理工具的自定义pypi镜像源配置
[[tool.poetry.source]]
name = "aliyun"
url = "http://mirrors.aliyun.com/pypi/simple"
default = false
secondary = true
